refactor: extract sandbox interpreter tests
